evaluation_tool/requirements.txt
pillow~=9.4.0
opencv
